fix(components): change config structure and fix seo
