+++ src/config/theme/index.ts
@@ -1,0 +1,2 @@
+export * from "./default";
+export * from "./osmosis";
